requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
bs4
tiktoken
re
request